refactor(data): give the four relational entity plans one owner
The plan that maps an entity onto a table was written out once per relational
adapter. The four copies had already drifted with nothing able to see it:
MySQL cached decoded roots under a case-insensitive comparer for no reason,
and three adapters assembled the same readiness target from three
differently-named properties.

RelationalEntityPlan owns the shared work. A store declares only what makes
it different — what qualifies its table name, how a root appears in a SELECT,
and whether a scalar is encoded on its way into a parameter. Schema and
Database disappear as separate concepts; the qualifier is one constructor
argument and SQLite's absence of one stops being a special case. The adapter
side goes from 642 lines to 142.

Two things the load-bearing check found, documented in place rather than
changed here, because a refactor preserves behaviour: SQLite's SELECT-list
alias is redundant while one table is in scope, and its scalar-write encoding
reaches no DATA-0100 type the suite can exercise, which leaves open whether
the other three adapters are missing that encoding or nobody needs it.

Verified: sqlite 49/49, relational 21/21, postgres 28/28 against a real
container; solution builds 0 warnings, 0 errors.
