+++ src/Connectors/Data/SqlServer/Runtime/SqlServerRepository.cs
@@ -306,7 +306,7 @@ internal sealed class SqlServerRepository<TEntity, TKey> :
     private async Task Ready(CancellationToken ct)
     {
         var plan = Plan;
-        var target = $"{plan.Schema}/{plan.Table}/{plan.Mapping.Id}";
+        var target = plan.Target;
         if (_options.SourcePlan.UsesLegacyProvisioningReadiness)
         {
             await _readiness.Provision(_options.SourcePlan, target,
