+++ src/Connectors/Data/Sqlite/Runtime/SqliteEntityPlan.cs
@@ -1,50 +1,28 @@
-using System.Linq.Expressions;
 using Koan.Data.Abstractions;
-using Koan.Data.Abstractions.Filtering;
 using Koan.Data.Core;
 using Koan.Data.Core.Semantics;
 using Koan.Data.Relational;
-using Koan.Data.Relational.Mapping;
 using Microsoft.Data.Sqlite;
 
 namespace Koan.Data.Connector.Sqlite.Runtime;
 
-internal sealed class SqliteEntityPlan<TEntity, TKey>
+/// <summary>
+/// SQLite has no qualifier above the table. It differs from its siblings in three places: reads address columns
+/// through a named subquery, scalars are written in DATA-0100's order-preserving form, and the client hands
+/// back a reader rather than a row.
+/// </summary>
+internal sealed class SqliteEntityPlan<TEntity, TKey> : RelationalEntityPlan<TEntity, TKey, SqliteDialect>
     where TEntity : class, IEntity<TKey>
     where TKey : notnull
 {
-    private readonly RelationalStructuredValueCodec _structured;
-    private readonly MappingBindingPlan[] _readBindings;
-    private readonly Dictionary<string, bool> _structuredRoots;
-    private readonly Action<TEntity, object?>? _generatedSetter;
-    private readonly MappingBindingPlan? _generatedBinding;
-
     internal SqliteEntityPlan(MappingPlan mapping, DataSegmentationPlan segmentation)
+        : base(mapping, segmentation, new SqliteDialect(), qualifier: null)
     {
         if (mapping.Container.Namespace.Count > 1 ||
             mapping.Container.Namespace.Count == 1 &&
             !string.Equals(mapping.Container.Namespace[0], "main", StringComparison.OrdinalIgnoreCase))
             throw new MappingCompilationException(mapping.Source, mapping.EntityType,
                 "SQLite mappings support the empty namespace or 'main'.");
-
-        Mapping = mapping;
-        Commands = new RelationalCommandPlanner(mapping);
-        Dialect = new SqliteDialect();
-        Table = mapping.Container.Name;
-        QualifiedTable = SqliteDialect.Quote(Table);
-        _readBindings = mapping.Read().Bindings.ToArray();
-        _structuredRoots = mapping.Bindings
-            .GroupBy(static binding => binding.PhysicalPath.Name, StringComparer.Ordinal)
-            .ToDictionary(
-                static group => group.Key,
-                static group => group.Any(binding =>
-                    binding.Shape == MappingValueShape.Object || binding.PhysicalPath.IsNested),
-                StringComparer.Ordinal);
-        _structured = new RelationalStructuredValueCodec(segmentation.For(typeof(TEntity)).Fields);
-        Roots = _readBindings.Select(static binding => binding.PhysicalPath.Name)
-            .Distinct(StringComparer.Ordinal).ToArray();
-        Select = string.Join(", ", Roots.Select(root => $"koan_row.{SqliteDialect.Quote(root)}"));
-        IdentityRoots = mapping.Identity.Parts.Select(static part => part.PhysicalPath.Name).ToArray();
 
         foreach (var identity in mapping.Identity.Parts)
         {
@@ -53,33 +31,30 @@ internal sealed class SqliteEntityPlan<TEntity, TKey>
                 throw new MappingCompilationException(mapping.Source, mapping.EntityType,
                     "SQLite identity components require scalar physical names.");
         }
-
-        if (mapping.Identity.IsGenerated && mapping.Identity.Parts.Count == 1 &&
-            mapping.Identity.LogicalPath.Segments.Count == 1)
-        {
-            _generatedBinding = mapping.Bindings.Single(binding => binding.Id == mapping.Identity.Parts[0].Id);
-            var property = typeof(TEntity).GetProperty(mapping.Identity.LogicalPath.Segments[0])
-                ?? throw new MappingValueException(mapping.Id, mapping.Identity.LogicalPath.ToString(),
-                    "Generated identity requires a writable Entity property.");
-            var setter = property.SetMethod
-                ?? throw new MappingValueException(mapping.Id, mapping.Identity.LogicalPath.ToString(),
-                    "Generated identity requires a writable Entity property.");
-            var entity = Expression.Parameter(typeof(TEntity), "entity");
-            var value = Expression.Parameter(typeof(object), "value");
-            _generatedSetter = Expression.Lambda<Action<TEntity, object?>>(
-                Expression.Call(entity, setter, Expression.Convert(value, property.PropertyType)), entity, value).Compile();
-        }
     }
 
-    internal MappingPlan Mapping { get; }
-    internal RelationalCommandPlanner Commands { get; }
-    internal SqliteDialect Dialect { get; }
-    internal string Table { get; }
-    internal string QualifiedTable { get; }
-    internal string Select { get; }
-    internal IReadOnlyList<string> Roots { get; }
-    internal IReadOnlyList<string> IdentityRoots { get; }
+    /// <summary>
+    /// Every statement this adapter writes aliases the table <c>koan_row</c> and addresses columns through it,
+    /// including the expressions <see cref="SqliteDialect.Read"/> produces. A SELECT list is the one place the
+    /// alias is redundant — one table is in scope, so a bare column resolves either way, and the suite passes
+    /// without it. It is spelled here anyway so that every column reference in the adapter reads alike.
+    /// </summary>
+    protected override string Project(string root) => $"koan_row.{SqliteDialect.Quote(root)}";
 
+    /// <summary>
+    /// A scalar written to a real column, in the order-preserving form DATA-0100 defines, so that a filter —
+    /// whose comparand is encoded the same way by <see cref="ComparableScalarEncoding.EncodeComparand"/> —
+    /// compares like for like.
+    ///
+    /// <para>Of the four relational adapters only this one encodes here; the other three write the value as it
+    /// comes. Disabling this leaves the suite green, including the DATA-0100 oracle, which says the branch
+    /// carries no governed type today — a document column is encoded by the codec's converters instead, and a
+    /// projected column is computed by the store. Whether that makes this defensive or makes the other three
+    /// wrong is an open question rather than something this seam settles.</para>
+    /// </summary>
+    protected override object? EncodeScalar(object? value) => ComparableScalarEncoding.EncodeComparand(value);
+
+    /// <summary>The client exposes a reader rather than a row, so the columns are lifted out of it first.</summary>
     internal TEntity Hydrate(SqliteDataReader reader)
     {
         var roots = new Dictionary<string, object?>(StringComparer.Ordinal);
@@ -88,88 +63,6 @@ internal sealed class SqliteEntityPlan<TEntity, TKey>
             var ordinal = reader.GetOrdinal(root);
             roots[root] = reader.IsDBNull(ordinal) ? null : reader.GetValue(ordinal);
         }
-
-        var decoded = new Dictionary<string, object?>(StringComparer.Ordinal);
-        var values = new List<MappedValue>(_readBindings.Length);
-        foreach (var binding in _readBindings)
-        {
-            if (!roots.TryGetValue(binding.PhysicalPath.Name, out var raw)) continue;
-            object? value;
-            if (_structuredRoots[binding.PhysicalPath.Name])
-            {
-                if (!decoded.TryGetValue(binding.PhysicalPath.Name, out var root))
-                {
-                    root = _structured.Deserialize(raw);
-                    decoded.Add(binding.PhysicalPath.Name, root);
-                }
-                value = binding.PhysicalPath.IsNested
-                    ? RelationalStructuredValueCodec.ReadPath(root, binding.PhysicalPath.Segments)
-                    : root;
-            }
-            else value = raw;
-            values.Add(new MappedValue(binding.Id, binding.PhysicalPath, binding.Shape, value));
-        }
-        return Mapping.Hydrate<TEntity>(values);
-    }
-
-    internal object? Parameter(RelationalValue value) => value.Binding.Shape == MappingValueShape.Object
-        ? _structured.Serialize(value.Value, value.Binding.LogicalPath.IsRoot)
-        : ComparableScalarEncoding.EncodeComparand(value.Value);
-
-    internal string JsonParameter(RelationalValue value) =>
-        _structured.Serialize(value.Value, includeManagedFields: false);
-
-    internal string NestedRoot(IEnumerable<RelationalValue> values)
-    {
-        var root = new Node();
-        foreach (var value in values)
-        {
-            var current = root;
-            foreach (var segment in value.Binding.PhysicalPath.Segments)
-            {
-                if (!current.Children.TryGetValue(segment, out var child))
-                {
-                    child = new Node();
-                    current.Children.Add(segment, child);
-                }
-                current = child;
-            }
-            current.Value = value.Value;
-            current.HasValue = true;
-        }
-        return _structured.Serialize(ToObject(root));
-    }
-
-    internal string ManagedPath(ResolvedField field) => ManagedPath(
-        field.StorageName ?? throw new MappingValueException(Mapping.Id, "managed", "Managed storage name is missing."),
-        field.ComparableType);
-
-    internal string ManagedPath(string storageName, Type type)
-    {
-        var root = Mapping.Bindings.SingleOrDefault(static binding =>
-            binding.Shape == MappingValueShape.Object && binding.LogicalPath.IsRoot)
-            ?? throw new MappingValueException(Mapping.Id, storageName,
-                "Declare a root Object binding to persist framework-managed fields.");
-        return Dialect.Read(new PhysicalPath(root.PhysicalPath.Name, storageName), MappingValueShape.Scalar, type);
-    }
-
-    internal bool IsStructuredRoot(string root) => _structuredRoots.TryGetValue(root, out var value) && value;
-
-    internal void AssignGenerated(TEntity entity, object? value)
-    {
-        if (_generatedSetter is null || _generatedBinding is null)
-            throw new MappingValueException(Mapping.Id, Mapping.Identity.LogicalPath.ToString(),
-                "This map does not declare one writable generated identity.");
-        _generatedSetter(entity, _generatedBinding.Decode(value));
-    }
-
-    private static DataObject ToObject(Node node) => new(node.Children.Select(pair =>
-        new DataProperty(pair.Key, pair.Value.HasValue ? pair.Value.Value : ToObject(pair.Value))));
-
-    private sealed class Node
-    {
-        internal Dictionary<string, Node> Children { get; } = new(StringComparer.Ordinal);
-        internal object? Value { get; set; }
-        internal bool HasValue { get; set; }
+        return Hydrate(roots);
     }
 }
